Skeleton for writing logical plan rewriting.
