+++ .scalafmt.conf
@@ -1,4 +1,4 @@
-maxColumn = 120
+maxColumn = 100
 continuationIndent.defnSite = 2
 binPack.parentConstructors = true
 newlines.penalizeSingleSelectMultiArgList = false
