+++ datasource/src/main/scala/astraea/spark/rasterframes/datasource/GeoTrellisDataSource.scala
@@ -1,7 +1,7 @@
 /*
  * This software is licensed under the Apache 2 license, quoted below.
  *
- * Copyright 2017 Azavea
+ * Copyright 2017 Azavea & Astraea, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -28,6 +28,7 @@ import org.apache.spark.sql.sources._
 
 /**
  * @author echeipesh
+ * @author sfitch
  */
 @Experimental
 class GeoTrellisDataSource extends DataSourceRegister with RelationProvider {
