+++ datasource/src/main/scala/astraea/spark/rasterframes/datasource/GeoTrellisRelation.scala
@@ -1,7 +1,7 @@
 /*
  * This software is licensed under the Apache 2 license, quoted below.
  *
- * Copyright 2017 Azavea
+ * Copyright 2017 Azavea & Astraea, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -23,7 +23,7 @@ import java.net.URI
 
 import astraea.spark.rasterframes._
 import astraea.spark.rasterframes.util._
-import geotrellis.raster.{Tile, TileFeature}
+import geotrellis.raster.Tile
 import geotrellis.spark.io._
 import geotrellis.spark.{LayerId, SpatialKey, TileLayerMetadata, _}
 import geotrellis.util.LazyLogging
@@ -31,10 +31,10 @@ import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.catalyst.encoders.ExpressionEncoder
 import org.apache.spark.sql.gt.types.TileUDT
 import org.apache.spark.sql.sources._
-import org.apache.spark.sql.types.{Metadata, StructField, StructType}
+import org.apache.spark.sql.types._
 import org.apache.spark.sql.{Row, SQLContext}
-import spray.json.JsValue
 import spray.json.DefaultJsonProtocol._
+import spray.json.JsValue
 
 import scala.reflect.runtime.universe._
 
@@ -57,7 +57,6 @@ case class GeoTrellisRelation(sqlContext: SQLContext, uri: URI, layerId: LayerId
       }
       val tt = Class.forName(h.valueClass) match {
         case c if c.isAssignableFrom(classOf[Tile]) ⇒ typeOf[Tile]
-        case c if c.isAssignableFrom(classOf[TileFeature[_, _]]) ⇒ typeOf[TileFeature[_, _]]
         case c ⇒ throw new UnsupportedOperationException("Unsupported tile type " + c)
       }
       (kt, tt)
@@ -66,29 +65,26 @@ case class GeoTrellisRelation(sqlContext: SQLContext, uri: URI, layerId: LayerId
   override def schema: StructType = {
     val skSchema = ExpressionEncoder[SpatialKey]().schema
 
-    val metadata = attributes.readMetadata[JsValue](layerId) |>
-      (m ⇒ Metadata.fromJson(m.compactPrint))
+    val skMetadata = attributes.readMetadata[JsValue](layerId) |>
+      (m ⇒ Metadata.fromJson(m.compactPrint)) |>
+      (Metadata.empty.append.attachContext(_).tagSpatialKey.build)
 
     val keyFields = keyType match {
       case t if t =:= typeOf[SpaceTimeKey] ⇒
         val tkSchema = ExpressionEncoder[TemporalKey]().schema
+        val tkMetadata = Metadata.empty.append.tagTemporalKey.build
         List(
-          StructField(SPATIAL_KEY_COLUMN, skSchema, nullable = false, metadata),
-          StructField(TEMPORAL_KEY_COLUMN, tkSchema, nullable = false)
+          StructField(SPATIAL_KEY_COLUMN, skSchema, nullable = false, skMetadata),
+          StructField(TEMPORAL_KEY_COLUMN, tkSchema, nullable = false, tkMetadata)
         )
       case t if t =:= typeOf[SpatialKey] ⇒
         List(
-          StructField(SPATIAL_KEY_COLUMN, skSchema, nullable = false, metadata)
+          StructField(SPATIAL_KEY_COLUMN, skSchema, nullable = false, skMetadata)
         )
-      case c ⇒ throw new UnsupportedOperationException("Unsupported key type " + c)
     }
 
     val tileFields = tileClass match {
       case t if t =:= typeOf[Tile]  ⇒
-        List(
-          StructField(TILE_COLUMN, TileUDT, nullable = true)
-        )
-      case t if t =:= typeOf[TileFeature[_, _]] ⇒
         List(
           StructField(TILE_COLUMN, TileUDT, nullable = true)
         )
@@ -99,11 +95,10 @@ case class GeoTrellisRelation(sqlContext: SQLContext, uri: URI, layerId: LayerId
 
   /** Declare filter handling. */
   override def unhandledFilters(filters: Array[Filter]): Array[Filter] = {
-    filters
-//    filters.filter {
-//      case (_:IsNotNull | _:IsNull) => true
-//      case _ => false
-//    }
+    filters.filter {
+      case (_:IsNotNull | _:IsNull) => true
+      case _ => false
+    }
   }
 
   override def buildScan(requiredColumns: Array[String], filters: Array[Filter]): RDD[Row] = {
@@ -116,13 +111,13 @@ case class GeoTrellisRelation(sqlContext: SQLContext, uri: URI, layerId: LayerId
 
 
     keyType match {
-      case t if t =:= typeOf[SpaceTimeKey] ⇒
+      case k if k =:= typeOf[SpaceTimeKey] ⇒
         reader.query[SpaceTimeKey, Tile, TileLayerMetadata[SpaceTimeKey]](layerId)
           .result
           .map { case (stk: SpaceTimeKey, tile: Tile) ⇒
             Row(stk.spatialKey, stk.temporalKey, tile)
           }
-      case t if t =:= typeOf[SpatialKey] ⇒
+      case k if k =:= typeOf[SpatialKey] ⇒
         reader.query[SpatialKey, Tile, TileLayerMetadata[SpatialKey]](layerId)
           .result
           .map { case (sk: SpatialKey, tile: Tile) ⇒
@@ -135,6 +130,5 @@ case class GeoTrellisRelation(sqlContext: SQLContext, uri: URI, layerId: LayerId
   override def sizeInBytes = {
     super.sizeInBytes
   }
-
 }
